LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/path-objects/2d.path.isPointInPath.empty-expected.txt

2d.path.isPointInPath.empty

isPointInPath() works when there is no path


PASS isPointInPath() works when there is no path 
